Move MainCamera and Input classes to global package
